Fixed 404 Error file
PLEASE, USE A CDN IT MAKES IT SO MUCH EASIER (and better for the user)!

The favicon has to stay in the root folder and MUST BE called favicon.ico.

Image folder was not accessible, because it was outside of root.

Torch on 404 page had a strange alignment, fixed now.